requirements.txt
pbr!=0.7,<1.0,>=0.6
git-up==1.3.0
GitPython==1.0.1
termcolor==1.1.0
PyYAML==3.10
pygithub==1.26.0